Fix DenseConjunctionBulkScorer hang on NO_MORE_DOCS overflow
TestSortOptimization intermittently hung in the ReadAheadMatchAllDocsQuery path while scoring with DenseConjunctionBulkScorer. A no-timeout native reproduction stalled at 100% CPU, and sampling showed the process spinning inside DenseConjunctionBulkScorer.score().

The root cause was an Int overflow in the window upper-bound calculation. When the lead iterator had advanced to NO_MORE_DOCS, the Kotlin port computed windowBase + WINDOW_SIZE as Int before clamping to max. That overflowed negative, produced a non-progressing window, and let the outer scoring loop spin forever.

Fix the port to match upstream Lucene semantics by widening to Long before adding WINDOW_SIZE and only then clamping and converting back to Int. This preserves monotonic window advancement even when windowBase is NO_MORE_DOCS.

+++ core/src/commonMain/kotlin/org/gnit/lucenekmp/search/DenseConjunctionBulkScorer.kt
@@ -73,7 +73,7 @@ internal class DenseConjunctionBulkScorer(iterators: MutableList<DocIdSetIterato
             }
 
             val windowBase = lead?.docID() ?: windowMax
-            windowMax = min(max, windowBase + WINDOW_SIZE)
+            windowMax = min(max.toLong(), windowBase.toLong() + WINDOW_SIZE).toInt()
             if (windowMax > windowBase) {
                 scoreWindowUsingBitSet(collector, acceptDocs, iterators, windowBase, windowMax)
             }
